fix(threads): give the thread its own composer so attachments stay in it
